requirements.txt
torch>=2.1.0
numpy>=1.24.0
gymnasium>=0.29.0
pyyaml>=6.0
tensorboard>=2.14.0
pytest>=7.4.0
tqdm>=4.66.0
scipy>=1.11.0
